+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changes
 
+### 3.0.0
+
+Updated all dependencies to latest major versions. Updated to work with latest version of lab which no longer supports the done callback in tests, as such the minimum version of lab this package will work with is 15.0.0 and minimum node version is 8.9
+Removed deprecated methods ```standardContructorTest```, ```LabTesting.methodParameterTest```, ```LabTesting.functionParameterTest```
+
 ### 2.3.0
 
 Added ```destructuredConstructorTest```, ```functionDestructuredParameterTest```, ```methodDestructuredParameterTest``` to allow for testing where destructured parameters are used.